practice_4_DRF/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
flake8==6.1.0
gunicorn==21.2.0
idna==3.4
isodate==0.6.1
mccabe==0.7.0
packaging==23.2
Pillow==10.0.1
psycopg2==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1